Collapse TaskExecuter implementations
Consolidate the existing chain of TaskExecuter implementations into a single
TaskExecutor implementation. The previous chain of classes made it difficult to
reason about the flow of task execution without jumping around between
classes that otherwise executed linearly.
