+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/TaskExecutor.java
@@ -17,10 +17,12 @@ package org.gradle.api.internal.tasks;
 
 import org.gradle.api.internal.TaskInternal;
 
-public interface TaskExecuter {
+public interface TaskExecutor {
+
     /**
      * Executes the given task. If the task fails with an exception, the exception is packaged in the provided task
      * state.
      */
-    TaskExecuterResult execute(TaskInternal task, TaskStateInternal state, TaskExecutionContext context);
+    void execute(TaskInternal task, TaskStateInternal state, TaskExecutionContext context);
+
 }
